# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "KPI Laravel API",
        "description": "Multi-tenant HR KPI platform. Covers v1 (existing) and v2 (multi-tenant) endpoints.",
        "contact": {
            "email": "admin@system.local"
        },
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Local / Development"
        }
    ],
    "paths": {
...
        "/api/v2/kpi/assignments": {
            "get": {
                "tags": [
                    "KPI Assignments"
                ],
                "summary": "List assignments",
                "operationId": "7903cb387ea7493f122f3b2e8d0c13dc",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "KPI Assignments"
                ],
                "summary": "Create assignment",
                "operationId": "697bb12fb92b78f045bd541967f58120",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/kpi/assignments/{id}": {
            "get": {
                "tags": [
                    "KPI Assignments"
                ],
                "summary": "Get assignment details",
                "operationId": "e59a183e8fccbd9a6e8e852728d016fc",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/kpi/assignments/{id}/submit": {
            "post": {
                "tags": [
                    "KPI Assignments"
                ],
                "summary": "Employee submits results",
                "description": "Employee submits their KPI values.",
                "operationId": "7a74e21f7a6d325c9503bbf548872ed3",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/kpi/assignments/{id}/approve": {
            "post": {
                "tags": [
                    "KPI Assignments"
                ],
                "summary": "HR/Manager approves",
                "description": "Reviewer approves the KPI submission.",
                "operationId": "d3853d967d0389e0870073f55ffc0a8a",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/kpi/assignments/{id}/reject": {
            "post": {
                "tags": [
                    "KPI Assignments"
                ],
                "summary": "HR/Manager rejects",
                "description": "Reviewer rejects the KPI submission.",
                "operationId": "410d74981cd477d590b07d926fed8ea7",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/kpi/assignments/mine": {
            "get": {
                "tags": [
                    "KPI Assignments"
                ],
                "summary": "Get my own assignments",
                "description": "My assignments — for currently authenticated employee.",
                "operationId": "fa7ec662fe6b75818d7cb7608b111b8f",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Include: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}